Migrate diagram UI extension to injectable commands (#174)
