+++ client/packages/glsp-sprotty/src/base/di.config.ts
@@ -14,11 +14,11 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { ContainerModule } from "inversify";
-import { Tool, TYPES } from "sprotty/lib";
+import { configureCommand, Tool, TYPES } from "sprotty/lib";
 import "../../css/glsp-sprotty.css";
 import { GLSP_TYPES } from "../types";
 import { GLSPCommandStack, IReadonlyModelAccess } from "./command-stack";
-import { DiagramUIExtensionActionHandlerInitializer, DiagramUIExtensionRegistry } from "./diagram-ui-extension/diagram-ui-extension-registry";
+import { DiagramUIExtensionRegistry, HideDiagramUIExtensionCommand, ShowDiagramUIExtensionCommand } from "./diagram-ui-extension/diagram-ui-extension-registry";
 import { ModelUpdateActionInitializer, ModelUpdateObserverRegistry } from "./model/model-update-observer-registry";
 import { createToolFactory, ToolManagerActionHandler } from "./tool-manager/tool-manager-action-handler";
 
@@ -39,7 +39,8 @@ const defaultGLSPModule = new ContainerModule((bind, unbind, isBound, rebind) =>
 
     // DiagramUIExtension registry initialization ------------------------------------
     bind(GLSP_TYPES.DiagramUIExtensionRegistry).to(DiagramUIExtensionRegistry).inSingletonScope();
-    bind(TYPES.IActionHandlerInitializer).to(DiagramUIExtensionActionHandlerInitializer)
+    configureCommand({ bind, isBound }, ShowDiagramUIExtensionCommand);
+    configureCommand({ bind, isBound }, HideDiagramUIExtensionCommand);
 
     // Tool manager initialization ------------------------------------
