+++ client/packages/glsp-sprotty/src/base/self-initializing-action-handler.ts
@@ -1,0 +1,31 @@
+/********************************************************************************
+ * Copyright (c) 2019 EclipseSource and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the Eclipse
+ * Public License v. 2.0 are satisfied: GNU General Public License, version 2
+ * with the GNU Classpath Exception which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ ********************************************************************************/
+import { injectable } from "inversify";
+import { Action, ActionHandlerRegistry, IActionHandler, IActionHandlerInitializer, ICommand } from "sprotty/lib";
+
+/**
+* Convinience class for classes that implement both `IActionHandler` and `IActionHandlerInitializer`
+*/
+@injectable()
+export abstract class SelfInitializingActionHandler implements IActionHandler, IActionHandlerInitializer {
+
+    initialize(registry: ActionHandlerRegistry) {
+        this.handledActionKinds.forEach(kind => registry.register(kind, this))
+    }
+
+    abstract handle(action: Action): ICommand | Action | void
+    abstract handledActionKinds: string[]
+}
