+++ client/packages/glsp-sprotty/src/base/tool-manager/tool-manager-action-handler.ts
@@ -19,11 +19,13 @@ import { isSetOperationsAction, OperationKind, SetOperationsAction } from "../..
 import { EdgeCreationTool, NodeCreationTool } from "../../features/tools/creation-tool";
 import { MouseDeleteTool } from "../../features/tools/delete-tool";
 import { GLSP_TYPES } from "../../types";
-import { SelfInitializingActionHandler } from "../diagram-ui-extension/diagram-ui-extension-registry";
+import { SelfInitializingActionHandler } from "../self-initializing-action-handler";
 
+const UNDEFINED_TOOL_ID = "undefined-tool"
 
 @injectable()
 export class ToolManagerActionHandler extends SelfInitializingActionHandler {
+
     @inject(GLSP_TYPES.IToolFactory) readonly toolFactory: (operationKind: string) => Tool
     @inject(TYPES.IToolManager) readonly toolManager: ToolManager
 
@@ -34,6 +36,7 @@ export class ToolManagerActionHandler extends SelfInitializingActionHandler {
             this.configure(action)
         }
     }
+
     configure(action: SetOperationsAction): any {
         const configuredTools = action.operations.map(op => {
             const tool = this.toolFactory(op.operationKind)
@@ -43,7 +46,6 @@ export class ToolManagerActionHandler extends SelfInitializingActionHandler {
 
             return tool;
         }).filter(tool => tool.id !== UNDEFINED_TOOL_ID)
-
         this.toolManager.registerTools(...configuredTools)
     }
 }
@@ -55,7 +57,7 @@ export function isTypeAware(tool: Tool): tool is Tool & TypeAware {
 export interface TypeAware {
     elementTypeId: string;
 }
-const UNDEFINED_TOOL_ID = "undefined-tool"
+
 export function createToolFactory(): interfaces.FactoryCreator<Tool> {
     return (context: interfaces.Context) => {
         return (operationKind: string) => {
