+++ client/packages/glsp-sprotty/src/index.ts
@@ -18,6 +18,7 @@ export * from './base/command-stack';
 export * from './base/diagram-ui-extension/diagram-ui-extension';
 export * from './base/edit-config/edit-config';
 export * from './base/model/model-update-observer-registry';
+export * from './base/self-initializing-action-handler';
 export * from './base/tool-manager/tool-manager-action-handler';
 export * from './features/change-bounds/model';
 export * from './features/change-bounds/resize';
